requirements.txt
dj-database-url==0.5.0
Django==1.10.6
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.8.2
Pillow==5.3.0
psycopg2==2.7.5
whitenoise==4.1
